# lib/api-zod/src/generated/types/moderationActionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ModerationActionInputAction } from './moderationActionInputAction';

export interface ModerationActionInput {
  action: ModerationActionInputAction;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  reason: string;
  confirmed: true;
}