# lib/api-zod/src/generated/types/moderationActionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ModerationAuditEntry } from './moderationAuditEntry';

export interface ModerationActionResult {
  action: string;
  hidden: boolean;
  suspended: boolean;
  audit: ModerationAuditEntry;
}